refactor(go-service): split quantizedsliding/action.go by responsibility
Split 865-line action.go into 5 focused files by responsibility:
- types.go: Type declarations, constants, package-level variables
- handlers.go: Run() dispatcher, all handle* methods, state helpers
- overrides.go: Pipeline override builder functions
- ocr.go: OCR text extraction and hit box parsing
- normalize.go: Input normalization and math utilities

This addresses PR #1299 review feedback requesting file splitting for
maintainability. Zero logic changes - pure mechanical refactoring.

All 28 functions preserved with identical signatures and logic.
Build verification: go build ./... && go vet ./... pass.
