+++ agent/go-service/quantizedsliding/normalize.go
@@ -1,0 +1,83 @@
+package quantizedsliding
+
+import (
+	"fmt"
+	"strings"
+)
+
+func clampClickRepeat(repeat int) int {
+	if repeat < 0 {
+		return 0
+	}
+	if repeat > maxClickRepeat {
+		return maxClickRepeat
+	}
+
+	return repeat
+}
+
+func normalizeButton(btn any) ([]int, error) {
+	numbers, err := normalizeIntSlice(btn)
+	if err != nil {
+		return nil, err
+	}
+
+	switch len(numbers) {
+	case 2:
+		return []int{numbers[0], numbers[1], 1, 1}, nil
+	case 4:
+		return []int{numbers[0], numbers[1], numbers[2], numbers[3]}, nil
+	default:
+		return nil, fmt.Errorf("button must be [x,y] or [x,y,w,h], got len=%d", len(numbers))
+	}
+}
+
+func normalizeButtonParam(btn any) (buttonTarget, error) {
+	if template, ok := btn.(string); ok {
+		template = strings.TrimSpace(template)
+		if template == "" {
+			return buttonTarget{}, fmt.Errorf("button template must not be empty")
+		}
+
+		return buttonTarget{template: template}, nil
+	}
+
+	coordinates, err := normalizeButton(btn)
+	if err != nil {
+		return buttonTarget{}, err
+	}
+
+	return buttonTarget{coordinates: coordinates}, nil
+}
+
+func normalizeIntSlice(raw any) ([]int, error) {
+	switch v := raw.(type) {
+	case []int:
+		return append([]int(nil), v...), nil
+	case []float64:
+		result := make([]int, 0, len(v))
+		for _, item := range v {
+			result = append(result, int(item))
+		}
+		return result, nil
+	case []any:
+		result := make([]int, 0, len(v))
+		for _, item := range v {
+			num, ok := item.(float64)
+			if !ok {
+				return nil, fmt.Errorf("unsupported number type %T", item)
+			}
+			result = append(result, int(num))
+		}
+		return result, nil
+	default:
+		return nil, fmt.Errorf("unsupported button type %T", raw)
+	}
+}
+
+func centerPoint(rect []int) (int, int) {
+	if len(rect) < 4 {
+		return 0, 0
+	}
+	return rect[0] + rect[2]/2 + centerPointOffset[0], rect[1] + rect[3]/2 + centerPointOffset[1]
+}
